test: fold dataclass_empty into dataclass_construct; collapse closure_chain into closure_capture
dataclass_empty is subsumed by adding one empty-body dataclass
declaration to dataclass_construct. closure_capture is replaced by
its three-level generalisation (formerly closure_chain), keeping the
established name.
