--- test/module-level/well-formed/dataclass_empty.expected
@@ -1,1 +1,0 @@
-ok
