--- test/module-level/well-formed/dataclass_empty.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class C:
-    pass
-
-print("ok")
