+++ test/module-level/well-formed/scopes/closure_capture.py
@@ -1,6 +1,8 @@
-def f():
-    x = 6
-    def g():
-        return x
-    return g
-print(f()())
+def outer(x):
+    def middle(y):
+        def inner(z):
+            return x + y + z
+        return inner
+    return middle
+
+print(outer(1)(2)(3))
