--- test/module-level/well-formed/scopes/closure_chain.expected
@@ -1,1 +1,0 @@
-6
